Move response heirarchy to its own file
